functional searchByName feature
